# openapi.yaml
openapi: 3.0.3
info:
  title: Arcadia Finance API
  version: "1.0.0"
  description: |
    REST API for the Arcadia Finance demo banking application.

    WARNING: This application is intentionally vulnerable (SQL injection,
    plain-text passwords, verbose errors) for F5 WAF/security demonstration
    purposes. NEVER deploy in production.

    ## Authentication

    All private endpoints require a valid JWT Bearer token.
    Tokens are issued by the app itself (HS256, 8-hour expiry).

    Workflow:
      1. POST /api/token  with username + password -> receive access_token
      2. Every request:   Authorization: Bearer <access_token>

  contact:
    name: Arcadia Finance Demo
  license:
    name: Internal Demo - Not for production use
servers:
  - url: http://localhost:8080
    description: Local Docker Compose stack
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /api/token or POST /api/login. Valid 8 hours.
  schemas:
...
    TransferRequest:
      type: object
      required: [from_account, to_account, amount]
      properties:
        from_account:
          type: string
          description: Source account number
          example: FR7601234001001
        to_account:
          type: string
          description: Destination account number
          example: FR7601234002001
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 250.00
        note:
          type: string
          description: Optional transfer memo
          example: "Invoice #42"
    TransferResponse:
      type: object
      properties:
        message:
          type: string
          example: Transfer completed successfully
        transfer_id:
          type: integer
          example: 14
        from_account:
          type: string
          example: FR7601234001001
        to_account:
          type: string
          example: FR7601234002001
        amount:
          type: number
          format: float
          example: 250.00
        note:
          type: string
          example: "Invoice #42"
    TransferRecord:
      type: object
      properties:
        id:
          type: number
          example: 7
        from_account:
          type: string
          example: FR7601234001001
        to_account:
          type: string
          example: FR7601234002001
        amount:
          type: number
          format: float
          example: 500.00
        note:
          type: string
          example: Dinner split
        status:
          type: string
          enum: [completed, pending, failed]
          example: completed
        created_at:
          type: string
          format: date-time
          example: "2026-07-01T10:30:00"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Not authenticated
        sql:
          type: string
          description: Intentional SQLi surface -- raw SQL returned in error responses
          example: "SELECT * FROM users WHERE username = '' OR '1'='1' --"
    StockQuote:
      type: object
      description: Normalised real-time quote returned by the stock-service.
      properties:
        ticker:
          type: string
          example: AAPL
        name:
          type: string
          example: Apple Inc.
        price:
          type: number
          format: float
          example: 213.49
        currency:
          type: string
          example: USD
        exchange:
          type: string
          example: NMS
        change:
          type: number
          format: float
          description: Price change since previous close
          example: 1.23
        change_pct:
          type: number
          format: float
          description: Percentage change since previous close
          example: 0.58
        market_cap:
          type: integer
          nullable: true
          example: 3280000000000
        pe_ratio:
          type: number
          format: float
          nullable: true
          example: 33.4
        volume:
          type: integer
          nullable: true
          example: 52341200
        day_high:
          type: number
          format: float
          nullable: true
          example: 215.10
        day_low:
          type: number
          format: float
          nullable: true
          example: 211.85
        fifty_two_week_high:
          type: number
          format: float
          nullable: true
          example: 237.23
        fifty_two_week_low:
          type: number
          format: float
          nullable: true
          example: 164.08
        sector:
          type: string
          nullable: true
          example: Technology
        industry:
          type: string
          nullable: true
          example: Consumer Electronics
    StockHistoryResponse:
      type: object
      properties:
        ticker:
          type: string
          example: AAPL
        period:
          type: string
          example: 1mo
        interval:
          type: string
          example: 1d
        data:
          type: array
          items:
            $ref: "#/components/schemas/OHLCVRecord"
    OHLCVRecord:
      type: object
      description: One OHLCV candle from the price history.
      properties:
        date:
          type: string
          format: date
          example: "2026-06-15"
        open:
          type: number
          format: float
          example: 210.50
        high:
          type: number
          format: float
          example: 214.30
        low:
          type: number
          format: float
          example: 209.80
        close:
          type: number
          format: float
          example: 213.49
        volume:
          type: integer
          example: 52341200
...
paths:
...
  /api/transfer:
    post:
      tags: [Transfers]
      summary: Execute a money transfer
      description: >
        Debits the source account and credits the destination account atomically.
        Proxied internally to the transfer-service (port 8081).
      operationId: createTransfer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
            example:
              from_account: FR7601234001001
              to_account: FR7601234002001
              amount: 250.00
              note: "Invoice #42"
      responses:
        "200":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Source or destination account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Transfer service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/transfers:
    get:
      tags: [Transfers]
      summary: Get transfer history for an account
      description: |
        Returns all transfers where the given account is sender or recipient,
        ordered by date descending (last 50 records).
        Proxied internally to the transfer-service (port 8081).
        NOTE: Intentionally vulnerable to SQL injection on the account param.
      operationId: listTransfers
      security:
        - BearerAuth: []
      parameters:
        - name: account
          in: query
          required: true
          description: Account number to retrieve history for
          schema:
            type: string
            example: FR7601234001001
      responses:
        "200":
          description: List of transfer records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferRecord"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Database error (verbose SQL returned -- intentional)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Transfer service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/stocks/quote:
    get:
      tags: [Stocks]
      summary: Get a live stock quote
      description: |
        Returns a real-time normalised quote for the requested ticker.
        Powered by the `get_stock_info` MCP tool (yfinance). Fields are
        normalised defensively: `currentPrice` / `regularMarketPrice` /
        `previousClose` for price; `shortName` / `longName` for name.
      operationId: getStockQuote
      security:
        - BearerAuth: []
      parameters:
        - name: ticker
          in: query
          required: true
          description: Stock ticker symbol (e.g. AAPL, MSFT, NVDA)
          schema:
            type: string
            example: AAPL
      responses:
        "200":
          description: Live quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockQuote"
              example:
                ticker: AAPL
                name: Apple Inc.
                price: 213.49
                currency: USD
                exchange: NMS
                change: 1.23
                change_pct: 0.58
                market_cap: 3280000000000
                pe_ratio: 33.4
                volume: 52341200
                day_high: 215.10
                day_low: 211.85
                fifty_two_week_high: 237.23
                fifty_two_week_low: 164.08
                sector: Technology
                industry: Consumer Electronics
        "400":
          description: Missing ticker parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Ticker not found on Yahoo Finance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: MCP / Yahoo Finance error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: stock-service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/stocks/history:
    get:
      tags: [Stocks]
      summary: Get historical OHLCV price data
      description: |
        Returns historical candles for a ticker. Powered by the
        `get_historical_stock_prices` MCP tool. Both `period` and `interval`
        are validated server-side against the yfinance allowed value sets.
      operationId: getStockHistory
      security:
        - BearerAuth: []
      parameters:
        - name: ticker
          in: query
          required: true
          schema:
            type: string
            example: AAPL
        - name: period
          in: query
          required: false
          description: "Lookback window. Allowed: 1d 5d 1mo 3mo 6mo 1y 2y 5y 10y ytd max"
          schema:
            type: string
            default: 1mo
            enum: [1d, 5d, 1mo, 3mo, 6mo, 1y, 2y, 5y, 10y, ytd, max]
            example: 1mo
        - name: interval
          in: query
          required: false
          description: "Candle interval. Allowed: 1m 2m 5m 15m 30m 60m 90m 1h 1d 5d 1wk 1mo 3mo"
          schema:
            type: string
            default: 1d
            enum: [1m, 2m, 5m, 15m, 30m, 60m, 90m, 1h, 1d, 5d, 1wk, 1mo, 3mo]
            example: 1d
      responses:
        "200":
          description: OHLCV records array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockHistoryResponse"
        "400":
          description: Missing or invalid parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Ticker not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: stock-service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/stocks/search:
    get:
      tags: [Stocks]
      summary: Validate / search a ticker symbol
      description: |
        Resolves the query string as a ticker via `get_stock_info`.
        Returns a short summary on success, 404 if not found.

        > **Note:** Yahoo Finance MCP has no company-name fuzzy search.
        > The query must be an exact ticker symbol (e.g. `AAPL`, not `Apple`).
      operationId: searchStock
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Ticker symbol to resolve (max 10 characters)
          schema:
            type: string
            maxLength: 10
            example: NVDA
      responses:
        "200":
          description: Ticker resolved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticker:   { type: string, example: NVDA }
                  name:     { type: string, example: NVIDIA Corporation }
                  price:    { type: number, format: float, example: 131.38 }
                  currency: { type: string, example: USD }
                  exchange: { type: string, example: NMS }
        "400":
          description: Missing or oversized query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Ticker not found on Yahoo Finance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: stock-service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...